Sample PHP files
dataenter.php, dataform.php and dataread.php are sample files I created
for a unit on simple use of PHP to connect to a MySQL database.  The
field names should be self-explanatory and simulate a simple contacts
database, which needs to be setup in MySQL ahead of time.  The names of
the database, table and the authentication information all needs to be
setup to match that of the database.  The presentation includes
information about the database setuo on slide 16.  Note that this was
originally written for Uniform Server, not XAMPP, but the database info
is the same.